# src/notifying/routes/swagger/docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Notifying Service API",
    "description": "RESTful API for pub/sub messaging system with topic management, subscription handling, and message broadcasting capabilities.",
    "version": "1.0.15",
    "contact": {
      "name": "NooblyJS Core Team",
      "url": "https://github.com/nooblyjs/nooblyjs-core"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "/services/notifying/api",
      "description": "Notifying Service API"
    }
  ],
...
  "paths": {
...
    "/settings": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get service settings",
        "description": "Retrieves the current settings for the notifying service in the default instance.",
        "operationId": "getSettings",
        "responses": {
          "200": {
            "description": "Settings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Service settings object"
                }
              }
            }
          },
          "500": {
            "description": "Failed to retrieve settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Save service settings",
        "description": "Saves or updates the settings for the notifying service in the default instance.",
        "operationId": "saveSettings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Settings object to save"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings saved successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "OK"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing settings",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Bad Request: Missing settings"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication"
      }
    }
  },
  "security": [
    {
      "ApiKey": []
    },
    {
      "BearerAuth": []
    }
  ]
}